Updated iOS project and compilation fixes.
git-svn-id: http://svn.purei.org/palleon/trunk@94 e658db38-2ae6-4b1c-8a17-0a8554443f32

+++ src/Emitter.cpp
@@ -72,7 +72,7 @@ void CEmitter::Emit(unsigned int emitCount)
 		{
 		case EMITTER_SOURCE_POINT:
 			{
-				float angle = RangeRandom(0.0f, 2.0f * M_PI);
+				float angle = RangeRandom(0.0f, 2.0f * static_cast<float>(M_PI));
 				particle.position = CVector2(0, 0);
 				particle.velocity = CVector2(sin(angle), cos(angle)) * speed;
 			}
